service-workers/service-worker/import-installed-module-scripts.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ServiceWorker: Import installed module scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
promise_test(async t => {
  const script = '/workers/modules/resources/static-import-worker.js';
  const type = 'module';

  await service_worker_unregister(t, script);
  await navigator.serviceWorker.register(script, {
    scope: script,
    type: type
  });
  const msgPromise = new Promise(resolve => {
    navigator.serviceWorker.onmessage = resolve;
  });
  const registration = await navigator.serviceWorker.register(script, {
    scope: script,
    type: type
  });

  // The timing of an activation of ServiceWorker is flaky.
  const sw = registration.installing ?
    registration.installing : registration.active;
  sw.postMessage('');
  const msgEvent = await msgPromise;
  assert_array_equals(msgEvent.data, ['export-on-load-script.js']);
}, 'Import installed module script');
</script>
</body>
